Added DynaTerrain
*Node tree automatically manages itself when a subdivision or refinement
occurs (handles automatically 1-level-max-of-difference rule between
neighbor nodes).

*The elevation data is handled by NzHeightSource class. For now only a
procedural function can be used for the height (e.g. a sine function, a
constant value, a NzNoise value, etc.) but reading elevation data file is
planned. The user needs to inherit NzHeightSource and redefine GetNoiseValue (btw the function's name will be changed to GetProceduralValue). The inherited class should then be passed along to the quadtree class (i.e. the terrain)

*The user only needs to interact with NzQuadtree and NzHeightSource to get
a full control of the terrain. In a server/client architecture, the
NzHeightSource class can be used server-side, as it has nothing to do with
rendering, it just provides height of any point of the terrain