Added Dispatcher & Zone class to deal with the GPU
*Dispatcher will receive all vertice buffers coming from Patches and
dispatch them towards the appropriate Zone. Also, Dispatcher centralizes
all the required "GPU-memory" (CPU copy of the GPU memory to be precise)
and splits big 1Mio buffers between individual Zones.

*Zone is at the same time a spatial area and a memory "conteneur". It does
not have owernship over the allocated memory (i.e. 1Mio buffers) but it
does all kinds of operations with it. Because different area of the
terrain have different needs in term of precision & so in triangle count
and memory, each Zone can have a different number of buffers in charge.
The Zone class will be efficient for culling because you can discard at
the same time many buffers for rendering. The size of a zone should have
similar dimensions to the camera frustum to ensure efficient and fast view
frustum culling